requirements.txt
anyio==3.6.2
appdirs==1.4.4
argcomplete==3.0.8
arrow==1.2.3
asgiref==3.5.2
attrs==22.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
binaryornot==0.4.4
blinker==1.5
certifi==2022.6.15
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
click-log==0.4.0
cloudinary==1.31.0
cloudinary-cli==1.5.1
contextlib2==21.6.0
cookiecutter==2.1.1
distlib==0.3.6
Django==4.1.3
django-bootstrap==0.2.4
django-bootstrap4==23.1
django-browser-reload==1.8.0
django-dropbox==0.0.2
django-filter==22.1
django-mailjet==0.3.1
django-storages==1.13.2
django-tailwind==3.5.0
docstring-parser==0.15
dropbox==11.36.0
ecdsa==0.18.0
exceptiongroup==1.1.0
fastapi==0.89.1
filelock==3.8.0
flacon==0.5.0
Flask==2.2.2
Flask-MySQLdb==1.0.1
fontawesomefree==6.2.1
functions==0.7.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jinja2-time==0.2.0
jose==1.0.0
mailjet-rest==1.3.4
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysqlclient==2.1.1
packaging==23.0
passlib==1.7.4
Pillow==9.5.0
pipenv==2022.11.30
platformdirs==2.5.4
pluggy==1.0.0
ply==3.11
prometheus-client==0.15.0
prometheus-flask-exporter==0.21.0
psycopg2==2.9.5
pyasn1==0.4.8
pydantic==1.10.4
Pygments==2.14.0
PyJWT==2.6.0
pypng==0.20220715.0
pyserial==3.5
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-escpos==2.2.0
python-jose==3.3.0
python-multipart==0.0.5
python-slugify==8.0.1
pytz==2023.3
pyusb==1.2.1
PyYAML==6.0
qrcode==7.4.2
raven==6.10.0
requests==2.28.2
rsa==4.9
schema==0.7.5
schemas==0.7.1
services==0.1.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.44
SQLAlchemy-Utils==0.39.0
sqlparse==0.4.3
starlette==0.22.0
stone==3.3.1
stripe==5.4.0
tailwindcss==0.0.1
text-unidecode==1.3
tk==0.1.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
virtualenv==20.16.7
virtualenv-clone==0.5.7
Werkzeug==2.2.2
whitenoise==6.4.0
